CORE-257 reset state when mounting ImportFlow.vue
